+++ modern.config.ts
@@ -1,12 +1,12 @@
-import { defineConfig, moduleTools } from '@modern-js/module-tools';
+import { defineConfig, moduleTools } from "@modern-js/module-tools";
 
 export default defineConfig({
   plugins: [moduleTools()],
   buildPreset: ({ extendPreset }) =>
-    extendPreset('npm-component', {
+    extendPreset("npm-component", {
       asset: {
         svgr: true,
-        path: '../assets',
+        path: "../assets",
       },
     }),
 });