+++ src/contributors/index.tsx
@@ -1,1 +1,1 @@
-export * from './contributors';
+export * from "./contributors";