+++ src/env.d.ts
@@ -1,9 +1,9 @@
-declare module '*.module.scss' {
+declare module "*.module.scss" {
   const classes: { readonly [key: string]: string };
   export default classes;
 }
 
-declare module '*.svg' {
+declare module "*.svg" {
   export const ReactComponent: React.FunctionComponent<React.SVGProps<SVGSVGElement>>;
 
   export default ReactComponent;