+++ src/popover/popover.tsx
@@ -1,7 +1,7 @@
-import Tooltip from 'rc-tooltip';
-import type { TooltipProps } from 'rc-tooltip/lib/Tooltip';
+import Tooltip from "rc-tooltip";
+import type { TooltipProps } from "rc-tooltip/lib/Tooltip";
 
-import './popover.scss';
+import "./popover.scss";
 
 export const Popover: React.FC<TooltipProps> = ({ ...rest }) => {
   return <Tooltip {...rest} />;