+++ src/theme/index.tsx
@@ -1,5 +1,5 @@
 import { Layout as DefaultLayout, HomeLayout, NotFoundLayout, useBindingAsideScroll, useSetup } from "rspress/theme";
-import { NavIcon } from "..";
+import { NavIcon } from "../nav-icon";
 
 const Layout = () => {
   return <DefaultLayout beforeNavTitle={<NavIcon />} />;